+++ apps/cli/CHANGELOG.md
@@ -1,4 +1,16 @@
-# @marshall/cli
+# @agentionai/marshall-cli
+
+## 0.3.0
+
+### Minor Changes
+
+- Rename from @marshall/_ to @agentionai/marshall-_
+
+### Patch Changes
+
+- Updated dependencies
+  - @agentionai/marshall-engine@0.3.0
+  - @agentionai/marshall-tools@0.3.0
 
 ## 0.2.0
 
@@ -9,5 +21,5 @@
 ### Patch Changes
 
 - Updated dependencies
-  - @marshall/engine@0.2.0
-  - @marshall/tools@0.2.0
+  - @agentionai/marshall-engine@0.2.0
+  - @agentionai/marshall-tools@0.2.0