+++ packages/engine/CHANGELOG.md
@@ -1,4 +1,15 @@
-# @marshall/engine
+# @agentionai/marshall-engine
+
+## 0.3.0
+
+### Minor Changes
+
+- Rename from @marshall/_ to @agentionai/marshall-_
+
+### Patch Changes
+
+- Updated dependencies
+  - @agentionai/marshall-tools@0.3.0
 
 ## 0.2.0
 
@@ -9,4 +20,4 @@
 ### Patch Changes
 
 - Updated dependencies
-  - @marshall/tools@0.2.0
+  - @agentionai/marshall-tools@0.2.0