+++ packages/tools/CHANGELOG.md
@@ -1,4 +1,10 @@
-# @marshall/tools
+# @agentionai/marshall-tools
+
+## 0.3.0
+
+### Minor Changes
+
+- Rename from @marshall/_ to @agention/marshall-_
 
 ## 0.2.0